refactor: split security groups across multiple files
