+++ src/constructs/ec2/security-groups/base.ts
@@ -1,6 +1,6 @@
 import type { CfnSecurityGroup, IPeer, SecurityGroupProps } from "@aws-cdk/aws-ec2";
 import { Peer, Port, SecurityGroup } from "@aws-cdk/aws-ec2";
-import type { GuStack } from "../core";
+import type { GuStack } from "../../core";
 
 /**
  * A way to describe an ingress or egress rule for a security group.
@@ -64,27 +64,6 @@ export class GuSecurityGroup extends SecurityGroup {
     props.egresses?.forEach(({ range, port, description }) => {
       const connection: Port = typeof port === "number" ? Port.tcp(port) : port;
       this.addEgressRule(range, connection, description);
-    });
-  }
-}
-
-export class GuWazuhAccess extends GuSecurityGroup {
-  private static getDefaultProps(): Partial<GuSecurityGroupProps> {
-    return {
-      description: "Wazuh agent registration and event logging",
-      overrideId: true,
-      allowAllOutbound: false,
-      egresses: [
-        { range: Peer.anyIpv4(), port: 1514, description: "Wazuh event logging" },
-        { range: Peer.anyIpv4(), port: 1515, description: "Wazuh agent registration" },
-      ],
-    };
-  }
-
-  constructor(scope: GuStack, id: string, props: GuSecurityGroupProps) {
-    super(scope, id, {
-      ...GuWazuhAccess.getDefaultProps(),
-      ...props,
     });
   }
 }
