+++ src/constructs/ec2/security-groups/index.ts
@@ -1,0 +1,2 @@
+export * from "./base";
+export * from "./wazuh";
